fix: worlds layout
